requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
sentence-transformers